Deduplicating further by parameterising on partial update callable.
