+++ flows/boundary.py
@@ -1,5 +1,6 @@
 """Boundary between Prefect, Vespa, and AWS."""
 
+import asyncio
 import base64
 import contextlib
 import json
@@ -9,7 +10,7 @@ import tempfile
 from collections.abc import Generator
 from io import BytesIO
 from pathlib import Path
-from typing import ContextManager, TypeAlias
+from typing import Callable, ContextManager, TypeAlias
 
 import boto3
 import vespa.querybuilder as qb
@@ -19,20 +20,29 @@ from cpr_sdk.models.search import Passage as VespaPassage
 from cpr_sdk.s3 import _get_s3_keys_with_prefix, _s3_object_read_text
 from cpr_sdk.search_adaptors import VespaSearchAdapter
 from cpr_sdk.ssm import get_aws_ssm_param
-from prefect import get_run_logger
+from prefect import flow, get_run_logger
+from prefect.deployments.deployments import run_deployment
 from prefect.logging import get_logger
 from pydantic import BaseModel
-from vespa.io import VespaQueryResponse
+from vespa.io import VespaQueryResponse, VespaResponse
 
-from scripts.cloud import ClassifierSpec
+from flows.utils import iterate_batch
+from scripts.cloud import (
+    AwsEnv,
+    ClassifierSpec,
+    function_to_flow_name,
+    generate_deployment_name,
+)
 from src.concept import Concept
-from src.exceptions import QueryError
+from src.exceptions import PartialUpdateError, QueryError
 from src.identifiers import WikibaseID
 from src.labelled_passage import LabelledPassage
 from src.span import Span
 
 HTTP_OK = 200
 CONCEPT_COUNT_SEPARATOR: str = ":"
+DEFAULT_DOCUMENTS_BATCH_SIZE = 500
+DEFAULT_INDEXING_TASK_BATCH_SIZE = 20
 
 # Needed to get document passages from Vespa
 # Example: CCLW.executive.1813.2418
@@ -683,3 +693,186 @@ class ConceptModel(BaseModel):
         raise ValueError(
             f"Could not extract concept name from model name '{self.model_name}'"
         )
+
+
+async def partial_update_text_block(
+    text_block_id: TextBlockId,
+    concepts: list[VespaConcept],  # A possibly empty list
+    document_import_id: DocumentImportId,
+    vespa_search_adapter: VespaSearchAdapter,
+    update_function: Callable,
+):
+    """Partial update a singular text block and its concepts."""
+    document_passage_id, document_passage = get_document_passage_from_vespa(
+        text_block_id, document_import_id, vespa_search_adapter
+    )
+
+    data_id = get_data_id_from_vespa_hit_id(document_passage_id)
+
+    serialised_concepts = update_function(document_passage, concepts)
+
+    response: VespaResponse = vespa_search_adapter.client.update_data(  # pyright: ignore[reportOptionalMemberAccess]
+        schema="document_passage",
+        namespace="doc_search",
+        data_id=data_id,
+        fields={"concepts": serialised_concepts},
+    )
+
+    if (status_code := response.get_status_code()) != HTTP_OK:
+        raise PartialUpdateError(data_id, status_code)
+
+    return None
+
+
+@flow
+async def run_partial_updates_of_concepts_for_batch(
+    documents_batch: list[DocumentImporter],
+    documents_batch_num: int,
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+    partial_update_flow: Callable,
+) -> None:
+    """Run partial updates for concepts in a batch of documents."""
+
+    logger = get_run_logger()
+    logger.info(
+        f"Updating concepts for batch of documents, documents in batch: {len(documents_batch)}."
+    )
+    for i, document_importer in enumerate(documents_batch):
+        try:
+            _ = await partial_update_flow(
+                document_importer=document_importer,
+                cache_bucket=cache_bucket,
+                concepts_counts_prefix=concepts_counts_prefix,
+            )
+
+            logger.info(f"processed batch documents #{documents_batch_num}")
+
+        except Exception as e:
+            document_stem: DocumentStem = documents_batch[i][0]
+            logger.error(
+                f"failed to process document `{document_stem}`: {e.__str__()}",
+            )
+            continue
+
+
+async def run_partial_updates_of_concepts_for_batch_flow_or_deployment(
+    documents_batch: list[DocumentImporter],
+    documents_batch_num: int,
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+    aws_env: AwsEnv,
+    as_deployment: bool,
+    partial_update_flow: Callable,
+) -> None:
+    """Run partial updates for a batch of documents as a sub-flow or deployment."""
+    logger = get_run_logger()
+    logger.info(
+        "Running partial updates of concepts for batch as sub-flow or deployment: "
+        f"batch length {len(documents_batch)}, as_deployment: {as_deployment}"
+    )
+
+    if as_deployment:
+        flow_name = function_to_flow_name(partial_update_flow)
+        deployment_name = generate_deployment_name(flow_name=flow_name, aws_env=aws_env)
+
+        return await run_deployment(
+            name=f"{flow_name}/{deployment_name}",
+            parameters={
+                "documents_batch": documents_batch,
+                "documents_batch_num": documents_batch_num,
+                "cache_bucket": cache_bucket,
+                "concepts_counts_prefix": concepts_counts_prefix,
+                "partial_update_flow": partial_update_flow,
+            },
+            timeout=3600,
+        )
+
+    return await run_partial_updates_of_concepts_for_batch(
+        documents_batch=documents_batch,
+        documents_batch_num=documents_batch_num,
+        cache_bucket=cache_bucket,
+        concepts_counts_prefix=concepts_counts_prefix,
+        partial_update_flow=partial_update_flow,
+    )
+
+
+@flow
+async def index_by_s3(
+    partial_update_flow: Callable,
+    aws_env: AwsEnv,
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+    vespa_search_adapter: VespaSearchAdapter | None,
+    s3_prefixes: list[str] | None = None,
+    s3_paths: list[str] | None = None,
+    batch_size: int = DEFAULT_DOCUMENTS_BATCH_SIZE,
+    indexing_task_batch_size: int = DEFAULT_INDEXING_TASK_BATCH_SIZE,
+    as_deployment: bool = True,
+) -> None:
+    """
+    Asynchronously index concepts from S3 files into Vespa.
+
+    This function retrieves concept documents from files stored in an S3 path and
+    indexes them in a Vespa instance. The name of each file in the specified S3 path is
+    expected to represent the document's import ID.
+
+    When `s3_prefix` is provided, the function will index all files within that S3
+    prefix (directory). When `s3_paths` is provided, the function will index only the
+    files specified in the list of S3 object keys. If both are provided `s3_paths` will
+    be used.
+
+    Assumptions:
+    - The S3 file names represent document import IDs.
+
+    params:
+    - s3_prefix: The S3 prefix (directory) to yield objects from.
+        E.g. "s3://bucket/prefix/"
+    - s3_paths: A list of S3 object keys to yield objects from.
+        E.g. {"s3://bucket/prefix/file1.json", "s3://bucket/prefix/file2.json"}
+    - vespa_search_adapter: An instance of VespaSearchAdapter.
+        E.g. VespaSearchAdapter(
+            instance_url="https://vespa-instance-url.com",
+            cert_directory="certs/"
+        )
+    """
+    logger = get_run_logger()
+
+    cm, vespa_search_adapter = get_vespa_search_adapter(vespa_search_adapter)
+
+    with cm:
+        logger.info("Getting S3 object generator")
+        documents_generator = s3_obj_generator(s3_prefixes, s3_paths)
+        documents_batches = iterate_batch(documents_generator, batch_size=batch_size)
+        indexing_task_batches = iterate_batch(
+            data=documents_batches, batch_size=indexing_task_batch_size
+        )
+
+        for i, indexing_task_batch in enumerate(indexing_task_batches, start=1):
+            logger.info(f"Processing indexing task batch #{i}")
+
+            indexing_tasks = [
+                run_partial_updates_of_concepts_for_batch_flow_or_deployment(
+                    documents_batch=documents_batch,
+                    documents_batch_num=documents_batch_num,
+                    cache_bucket=cache_bucket,
+                    concepts_counts_prefix=concepts_counts_prefix,
+                    aws_env=aws_env,
+                    as_deployment=as_deployment,
+                    partial_update_flow=partial_update_flow,
+                )
+                for documents_batch_num, documents_batch in enumerate(
+                    indexing_task_batch, start=1
+                )
+            ]
+
+            logger.info(f"Gathering indexing tasks for batch #{i}")
+            results = await asyncio.gather(*indexing_tasks, return_exceptions=True)
+            logger.info(f"Gathered indexing tasks for batch #{i}")
+
+            for result in results:
+                if isinstance(result, Exception):
+                    logger.error(
+                        f"failed to process document batch in indexing task batch #{i}: {str(result)}",
+                    )
+                    continue
