+++ tests/flows/test_index.py
@@ -13,16 +13,16 @@ from prefect.logging import disable_run_logger
 
 from flows.boundary import (
     ConceptModel,
+    get_document_passage_from_vespa,
     get_document_passages_from_vespa,
+    index_by_s3,
+    partial_update_text_block,
 )
 from flows.index import (
     CONCEPTS_COUNTS_PREFIX_DEFAULT,
     Config,
-    get_document_passage_from_vespa,
-    index_by_s3,
     index_labelled_passages_from_s3_to_vespa,
-    partial_update_text_block,
-    run_partial_updates_of_concepts_for_document_passages,
+    run_partial_updates_of_concepts_for_document_passages__update,
     update_concepts_on_existing_vespa_concepts,
 )
 from scripts.cloud import (
@@ -158,7 +158,7 @@ async def test_run_partial_updates_of_concepts_for_document_passages(
 
     assert (
         test_counts
-        == await run_partial_updates_of_concepts_for_document_passages.fn(
+        == await run_partial_updates_of_concepts_for_document_passages__update.fn(
             document_importer=(document_import_id, document_object_uri),
             vespa_search_adapter=local_vespa_search_adapter,
             cache_bucket=mock_bucket,
@@ -217,7 +217,7 @@ async def test_run_partial_updates_of_concepts_for_document_passages_task_failur
         # Run the update
         assert (
             Counter()
-            == await run_partial_updates_of_concepts_for_document_passages.fn(
+            == await run_partial_updates_of_concepts_for_document_passages__update.fn(
                 document_importer=(document_import_id, document_object_uri),
                 vespa_search_adapter=local_vespa_search_adapter,
                 cache_bucket=mock_bucket,
@@ -258,6 +258,7 @@ async def test_index_by_s3_with_s3_prefixes(
     )
 
     await index_by_s3(
+        partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
         aws_env=AwsEnv.sandbox,
         vespa_search_adapter=local_vespa_search_adapter,
         s3_prefixes=[os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)],
@@ -303,6 +304,7 @@ async def test_index_by_s3_with_s3_paths(
     ]
 
     await index_by_s3(
+        partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
         aws_env=AwsEnv.sandbox,
         vespa_search_adapter=local_vespa_search_adapter,
         s3_prefixes=None,
@@ -347,6 +349,7 @@ async def test_index_by_s3_task_failure(
         side_effect=mock_run_partial_updates_of_concepts_for_batch_flow_or_deployment,
     ):
         await index_by_s3(
+            partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
             aws_env=AwsEnv.sandbox,
             vespa_search_adapter=local_vespa_search_adapter,
             s3_prefixes=[
@@ -402,6 +405,7 @@ async def test_partial_update_text_block(
         [concept_a],
         document_import_id,
         local_vespa_search_adapter,
+        update_concepts_on_existing_vespa_concepts,
     )
 
     assert result_a is None
@@ -426,6 +430,7 @@ async def test_partial_update_text_block(
         [concept_b],
         document_import_id,
         local_vespa_search_adapter,
+        update_concepts_on_existing_vespa_concepts,
     )
 
     assert result_b is None
